Factor out some of the massiveness of the RequestorPage file
