+++ frontend/src/pages/RequestorPage/RequestorPage.tsx
@@ -1,12 +1,11 @@
 import { useToast } from "@/components/ui/use-toast";
 import { useWebsocketQueryInvalidation } from "@/hooks";
-import { cn, parsePathFromRequestUrl } from "@/utils";
+import { cn } from "@/utils";
 import { useCallback, useMemo, useRef } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
-import { KeyValueParameter, createKeyValueParameters } from "./KeyValueForm";
+import { KeyValueParameter } from "./KeyValueForm";
 import { RequestPanel } from "./RequestPanel";
 import { RequestorInput } from "./RequestorInput";
-import { useSessionHistory } from "./RequestorSessionHistoryContext";
 import { ResponsePanel } from "./ResponsePanel";
 import { RoutesCombobox } from "./RoutesCombobox";
 import { RoutesPanel } from "./RoutesPanel";
@@ -15,22 +14,22 @@ import {
   MakeProxiedRequestQueryFn,
   type ProbedRoute,
   Requestornator,
-  useFetchRequestorRequests,
   useMakeProxiedRequest,
-  // useMakeRequest,
 } from "./queries";
 import {
   type RequestorBody,
   type RequestorState,
   useRequestor,
 } from "./reducer";
-import { findMatchedRoute, useRoutes } from "./routes";
+import { useRoutes } from "./routes";
 import { BACKGROUND_LAYER } from "./styles";
-import { RequestMethodInputValue, isRequestMethod, isWsRequest } from "./types";
+import { isWsRequest } from "./types";
 import { useMakeWebsocketRequest } from "./useMakeWebsocketRequest";
 
 // We need some special CSS for grid layout that tailwind cannot handle
 import "./styles.css";
+import { useRequestorHistory } from "./useRequestorHistory";
+import { sortRequestornatorsDescending } from "./utils";
 
 export const RequestorPage = () => {
   const { toast } = useToast();
@@ -38,15 +37,9 @@ export const RequestorPage = () => {
   // Refresh routes in response to filesystem updates
   useWebsocketQueryInvalidation();
 
-  // TODO - Bring back persisted state once reducer is fully integrated
-  // const browserHistoryState = usePersistedUiState();
-
-  // ========================//
-  // === Refactored state ===//
-  // ========================//
-  const refactoredState = useRequestor();
-  // @ts-expect-error - testing
-  globalThis.refactoredState = refactoredState;
+  const requestorState = useRequestor();
+  // @ts-expect-error - This is helpful for debugging, soz
+  globalThis.refactoredState = requestorState;
   const {
     // Routes panel
     state: { routes },
@@ -55,7 +48,7 @@ export const RequestorPage = () => {
     getActiveRoute,
 
     // Requestor input
-    // - note that `requestType` is an internal property used to determine if we're making a websocket request or not
+    // NOTE - `requestType` is an internal property used to determine if we're making a websocket request or not
     state: { path, method, requestType },
     updatePath: handlePathInputChange,
     updateMethod: handleMethodChange,
@@ -84,7 +77,7 @@ export const RequestorPage = () => {
     state: { activeResponsePanelTab },
     setActiveResponsePanelTab,
     shouldShowResponseTab,
-  } = refactoredState;
+  } = requestorState;
 
   const selectedRoute = getActiveRoute();
 
@@ -92,7 +85,7 @@ export const RequestorPage = () => {
     setRoutes,
   });
 
-  // NOTE - Use this to test overflow
+  // NOTE - Use this to test overflow of requests panel
   // useEffect(() => {
   //   setQueryParams(
   //     createKeyValueParameters(
@@ -151,6 +144,7 @@ export const RequestorPage = () => {
 
   const formRef = useRef<HTMLFormElement>(null);
 
+  // FIXME / INVESTIGATE - Should this behavior change for websockets?
   useHotkeys(
     "mod+enter",
     () => {
@@ -243,6 +237,7 @@ export const RequestorPage = () => {
       <div
         className={cn(
           "grid",
+          // This is a custom css class that uses the famed `auto minmax(0, 1fr)` trick
           "fpx-requestor-grid-rows",
           "gap-2",
           // HACK - This is a workaround to prevent the grid from overflowing on smaller screens
@@ -275,8 +270,8 @@ export const RequestorPage = () => {
             "sm:grid-cols-[auto_1fr]",
             "rounded-md",
             "border",
-            // HACK - This prevents overflow from getting too excessive.
-            // FIXME - Need to resolve the problem with inner content expanding the parent
+            // HACK - This defensively prevents overflow from getting too excessive,
+            //        In the case where the inner content expands beyond the parent
             "max-w-screen",
             "max-h-full",
           )}
@@ -327,170 +322,6 @@ export const RequestorPage = () => {
 };
 
 export default RequestorPage;
-
-type RequestorHistoryHookArgs = {
-  routes: ProbedRoute[];
-  handleSelectRoute: (r: ProbedRoute, pathParams?: KeyValueParameter[]) => void;
-  setPath: (path: string) => void;
-  setMethod: (method: RequestMethodInputValue) => void;
-  setBody: (body: string | undefined) => void;
-  setPathParams: (headers: KeyValueParameter[]) => void;
-  setQueryParams: (params: KeyValueParameter[]) => void;
-  setRequestHeaders: (headers: KeyValueParameter[]) => void;
-};
-
-function useRequestorHistory({
-  routes,
-  handleSelectRoute,
-  setPath,
-  setMethod,
-  setRequestHeaders,
-  setBody,
-  setQueryParams,
-}: RequestorHistoryHookArgs) {
-  const {
-    sessionHistory: sessionHistoryTraceIds,
-    recordRequestInSessionHistory,
-  } = useSessionHistory();
-  const { data: allRequests } = useFetchRequestorRequests();
-
-  // Keep a history of recent requests and responses
-  const history = useMemo<Array<Requestornator>>(() => {
-    if (allRequests) {
-      const cloned = [...allRequests];
-      cloned.sort(sortRequestornatorsDescending);
-      return cloned;
-    }
-    return [];
-  }, [allRequests]);
-
-  // This feels wrong... but it's a way to load a past request back into the UI
-  const loadHistoricalRequest = (traceId: string) => {
-    recordRequestInSessionHistory(traceId);
-    const match = history.find((r) => r.app_responses?.traceId === traceId);
-    if (match) {
-      const method = match.app_requests.requestMethod;
-      let routePattern = match.app_requests.requestRoute;
-      // HACK - In case it's an unqualified route
-      if (routePattern === "") {
-        routePattern = "/";
-      }
-      const requestType = match.app_requests.requestUrl.startsWith("ws")
-        ? "websocket"
-        : "http";
-      const matchedRoute = findMatchedRoute(
-        routes,
-        routePattern,
-        method,
-        requestType,
-      );
-
-      if (matchedRoute) {
-        const pathParamsObject = match.app_requests.requestPathParams ?? {};
-        const pathParams = createKeyValueParameters(
-          Object.entries(pathParamsObject).map(([key, value]) => ({
-            key,
-            value,
-          })),
-        );
-
-        // NOTE - Helps us set path parameters correctly
-        handleSelectRoute(matchedRoute.route, pathParams);
-
-        // Reset the path to the *exact* path of the request, instead of the route pattern
-        const path =
-          parsePathFromRequestUrl(match.app_requests.requestUrl) ?? "";
-        setPath(path);
-
-        const headers = match.app_requests.requestHeaders ?? {};
-        setRequestHeaders(
-          createKeyValueParameters(
-            Object.entries(headers).map(([key, value]) => ({ key, value })),
-          ),
-        );
-
-        const queryParams = match.app_requests.requestQueryParams ?? {};
-        setQueryParams(
-          createKeyValueParameters(
-            Object.entries(queryParams).map(([key, value]) => ({
-              key,
-              value,
-            })),
-          ),
-        );
-
-        // NOTE - We set the body to be undefined or a (json serialized) string for now,
-        //        since that helps us render it in the UI (specifically in CodeMirror editors)
-        const body = match.app_requests.requestBody;
-        if (body === undefined || body === null) {
-          setBody(undefined);
-        } else {
-          const safeBody =
-            typeof body !== "string" ? JSON.stringify(body) : body;
-          setBody(safeBody);
-        }
-      } else {
-        // HACK - move this logic into the reducer
-        // Reset the path to the *exact* path of the request, instead of the route pattern
-        const path =
-          parsePathFromRequestUrl(match.app_requests.requestUrl) ?? "";
-        setPath(path);
-
-        const requestType = match.app_requests.requestUrl.startsWith("ws")
-          ? "websocket"
-          : "http";
-
-        setMethod(
-          isWsRequest(requestType)
-            ? "WS"
-            : isRequestMethod(method)
-              ? method
-              : "GET",
-        );
-
-        const headers = match.app_requests.requestHeaders ?? {};
-        setRequestHeaders(
-          createKeyValueParameters(
-            Object.entries(headers).map(([key, value]) => ({ key, value })),
-          ),
-        );
-
-        const queryParams = match.app_requests.requestQueryParams ?? {};
-        setQueryParams(
-          createKeyValueParameters(
-            Object.entries(queryParams).map(([key, value]) => ({
-              key,
-              value,
-            })),
-          ),
-        );
-      }
-    }
-  };
-
-  // Keep a local history of requests that the user has made in the UI
-  // This should be a subset of the full history
-  // These will be cleared on page reload
-  const sessionHistory = useMemo(() => {
-    return sessionHistoryTraceIds.reduce(
-      (matchedRequestornators, traceId) => {
-        const match = history.find((r) => r.app_responses?.traceId === traceId);
-        if (match) {
-          matchedRequestornators.push(match);
-        }
-        return matchedRequestornators;
-      },
-      [] as Array<Requestornator>,
-    );
-  }, [history, sessionHistoryTraceIds]);
-
-  return {
-    history,
-    sessionHistory,
-    recordRequestInSessionHistory,
-    loadHistoricalRequest,
-  };
-}
 
 function useRequestorSubmitHandler({
   requestType,
@@ -640,6 +471,15 @@ function getContentTypeHeader(body: RequestorBody): string | null {
   }
 }
 
+function getContentLength(body: RequestorBody) {
+  switch (body.type) {
+    case "file":
+      return body.value?.size ?? null;
+    default:
+      return null;
+  }
+}
+
 /**
  * When you select a route from the route side panel,
  * this will look for the most recent request made against that route.
@@ -660,25 +500,4 @@ function useMostRecentRequestornator(
 
     return matchingResponses?.[0];
   }, [all, requestInputs]);
-}
-
-function sortRequestornatorsDescending(a: Requestornator, b: Requestornator) {
-  const aLatestTimestamp = a.app_requests?.updatedAt;
-  const bLatestTimestamp = b.app_requests?.updatedAt;
-  if (aLatestTimestamp > bLatestTimestamp) {
-    return -1;
-  }
-  if (aLatestTimestamp < bLatestTimestamp) {
-    return 1;
-  }
-  return 0;
-}
-
-function getContentLength(body: RequestorBody) {
-  switch (body.type) {
-    case "file":
-      return body.value?.size ?? null;
-    default:
-      return null;
-  }
 }
