+++ frontend/src/pages/RequestorPage/useRequestorHistory.ts
@@ -1,0 +1,176 @@
+import { parsePathFromRequestUrl } from "@/utils";
+import { useMemo } from "react";
+import { KeyValueParameter, createKeyValueParameters } from "./KeyValueForm";
+import { useSessionHistory } from "./RequestorSessionHistoryContext";
+import {
+  ProbedRoute,
+  Requestornator,
+  useFetchRequestorRequests,
+} from "./queries";
+import { findMatchedRoute } from "./routes";
+import { RequestMethodInputValue, isRequestMethod, isWsRequest } from "./types";
+import { sortRequestornatorsDescending } from "./utils";
+
+type RequestorHistoryHookArgs = {
+  routes: ProbedRoute[];
+  handleSelectRoute: (r: ProbedRoute, pathParams?: KeyValueParameter[]) => void;
+  setPath: (path: string) => void;
+  setMethod: (method: RequestMethodInputValue) => void;
+  setBody: (body: string | undefined) => void;
+  setPathParams: (headers: KeyValueParameter[]) => void;
+  setQueryParams: (params: KeyValueParameter[]) => void;
+  setRequestHeaders: (headers: KeyValueParameter[]) => void;
+};
+
+export function useRequestorHistory({
+  routes,
+  handleSelectRoute,
+  setPath,
+  setMethod,
+  setRequestHeaders,
+  setBody,
+  setQueryParams,
+}: RequestorHistoryHookArgs) {
+  const {
+    sessionHistory: sessionHistoryTraceIds,
+    recordRequestInSessionHistory,
+  } = useSessionHistory();
+  const { data: allRequests } = useFetchRequestorRequests();
+
+  // Keep a history of recent requests and responses
+  const history = useMemo<Array<Requestornator>>(() => {
+    if (allRequests) {
+      const cloned = [...allRequests];
+      cloned.sort(sortRequestornatorsDescending);
+      return cloned;
+    }
+    return [];
+  }, [allRequests]);
+
+  // This feels wrong... but it's a way to load a past request back into the UI
+  const loadHistoricalRequest = (traceId: string) => {
+    recordRequestInSessionHistory(traceId);
+    const match = history.find((r) => r.app_responses?.traceId === traceId);
+    if (match) {
+      const method = match.app_requests.requestMethod;
+      let routePattern = match.app_requests.requestRoute;
+      // HACK - In case it's an unqualified route
+      if (routePattern === "") {
+        routePattern = "/";
+      }
+      const requestType = match.app_requests.requestUrl.startsWith("ws")
+        ? "websocket"
+        : "http";
+      const matchedRoute = findMatchedRoute(
+        routes,
+        routePattern,
+        method,
+        requestType,
+      );
+
+      if (matchedRoute) {
+        const pathParamsObject = match.app_requests.requestPathParams ?? {};
+        const pathParams = createKeyValueParameters(
+          Object.entries(pathParamsObject).map(([key, value]) => ({
+            key,
+            value,
+          })),
+        );
+
+        // NOTE - Helps us set path parameters correctly
+        handleSelectRoute(matchedRoute.route, pathParams);
+
+        // Reset the path to the *exact* path of the request, instead of the route pattern
+        const path =
+          parsePathFromRequestUrl(match.app_requests.requestUrl) ?? "";
+        setPath(path);
+
+        const headers = match.app_requests.requestHeaders ?? {};
+        setRequestHeaders(
+          createKeyValueParameters(
+            Object.entries(headers).map(([key, value]) => ({ key, value })),
+          ),
+        );
+
+        const queryParams = match.app_requests.requestQueryParams ?? {};
+        setQueryParams(
+          createKeyValueParameters(
+            Object.entries(queryParams).map(([key, value]) => ({
+              key,
+              value,
+            })),
+          ),
+        );
+
+        // NOTE - We set the body to be undefined or a (json serialized) string for now,
+        //        since that helps us render it in the UI (specifically in CodeMirror editors)
+        const body = match.app_requests.requestBody;
+        if (body === undefined || body === null) {
+          setBody(undefined);
+        } else {
+          const safeBody =
+            typeof body !== "string" ? JSON.stringify(body) : body;
+          setBody(safeBody);
+        }
+      } else {
+        // HACK - move this logic into the reducer
+        // Reset the path to the *exact* path of the request, instead of the route pattern
+        const path =
+          parsePathFromRequestUrl(match.app_requests.requestUrl) ?? "";
+        setPath(path);
+
+        const requestType = match.app_requests.requestUrl.startsWith("ws")
+          ? "websocket"
+          : "http";
+
+        setMethod(
+          isWsRequest(requestType)
+            ? "WS"
+            : isRequestMethod(method)
+              ? method
+              : "GET",
+        );
+
+        const headers = match.app_requests.requestHeaders ?? {};
+        setRequestHeaders(
+          createKeyValueParameters(
+            Object.entries(headers).map(([key, value]) => ({ key, value })),
+          ),
+        );
+
+        const queryParams = match.app_requests.requestQueryParams ?? {};
+        setQueryParams(
+          createKeyValueParameters(
+            Object.entries(queryParams).map(([key, value]) => ({
+              key,
+              value,
+            })),
+          ),
+        );
+      }
+    }
+  };
+
+  // Keep a local history of requests that the user has made in the UI
+  // This should be a subset of the full history
+  // These will be cleared on page reload
+  const sessionHistory = useMemo(() => {
+    return sessionHistoryTraceIds.reduce(
+      (matchedRequestornators, traceId) => {
+        const match = history.find((r) => r.app_responses?.traceId === traceId);
+        if (match) {
+          matchedRequestornators.push(match);
+        }
+        return matchedRequestornators;
+      },
+      [] as Array<Requestornator>,
+    );
+  }, [history, sessionHistoryTraceIds]);
+
+  return {
+    history,
+    sessionHistory,
+    recordRequestInSessionHistory,
+    loadHistoricalRequest,
+  };
+}
